Add files via upload
Really all in one minimalist web app framework for python 3.7

supports both MySQL and SQLite - no need for apache just point your browser to http://localhost:8000/

I got peeved with all those frameworks that promise wonders and deliver much less.

This is a port to Python 3.7 of a minimal web app I did in PHP some while back.

It really lives up to its name of "all in one" as this readme text and the css styling are all
tucked away in the allinOne.py file!

I tried to segregate my data structures from the executable code. The benefit in my view is that the data definitions exist ONLY in one place
 - the SQL statements to build the tables and views.

There is a way in both MySQL and SQLite (which are both supported in this project) to interogate
the schema and iterate thru the fields of a data entry form (ie the columns of a table).
Therefore you do not polute your data manager code with knowledge of the data structures. Any changes made
to these doesn't involve any rework - especially in forms and reports.

This comes at the cost of some begnin naming conventions. If you have a foreign key your your record you call it 
foreign_id and you expect the value to be displayed for this lookup field to be in a column called foreign in a 
table aptly called foreign_tb or a view called foreign_tb_view. Quite simple really.